refactor(actor): group src/ into actor/, mail/, wasm/ submodules
Folds the flat 12-file aether-actor src/ into three groups that mirror
the SDK's conceptual seams: actor/ for the marker traits + per-mail
ctx + sender + slot, mail/ for the inbound envelope + mailbox + sync
helper + transport trait, and wasm/ for the wasm-guest impl folded
together with its raw FFI submodule (eliminating the awkward
wasm.rs + wasm/raw.rs side-by-side layout). local.rs and log.rs stay
top-level as single cohesive concepts. Public API unchanged — every
crate-root re-export still resolves to the same type.
